Ensure existing swapfile is not disabled before enabling
Even though `swap_file_existing_size_mb` is converted to an integer on
its assignment, for some reason Ansible still sees it as a string. Hence
the comparison always fails, even when the swap file exists and already
has the correct size (e.g. because '2048' != 2048).

Add explicit conversion to integers in the comparison in order to make
sure that the correct values are compared.